create consolidated registries for eog andd fill first two layers
Former-commit-id: e74376cbcc9daeeb19d015d3ad529151e94a0370

+++ db/migrate/20140820171517_create_consolidated_registry.rb
@@ -1,6 +1,5 @@
 class CreateConsolidatedRegistry < ActiveRecord::Migration
   def self.up
-  #unless Project.name.to_sym == :eog
     # Start registry migration...
     create_table :registry_entries do |t|
       t.string :entry_code
@@ -78,16 +77,16 @@ class CreateConsolidatedRegistry < ActiveRecord::Migration
       t.timestamps
     end
 
-    create_table :languages do |t|
-      t.string :code
+    unless Project.name.to_sym == :eog
+      create_table :languages do |t|
+        t.string :code
+      end
+
+      add_column :interviews, :language_id, :integer
+      drop_table :categorizations
+      drop_table :category_translations
+      drop_table :categories
     end
-
-    add_column :interviews, :language_id, :integer
-
-    drop_table :categorizations
-    drop_table :category_translations
-    drop_table :categories
-  #end
     Language.create_translation_table! :abbreviated => :string, :name => :string
   end
